# contracts/generated/billing-payments.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Otziv billing and public payment reads",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/api/common-billing/accounts/by-company/{companyId}": {
      "get": {
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CommonBillingAccountResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommonBillingAccountResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/common-billing/accounts/{accountId}": {
      "get": {
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CommonBillingAccountResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonBillingAccountResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommonBillingCompanyResponse": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "companyId",
          "companyTitle",
          "enabled"
        ],
        "properties": {
          "companyId": {
            "type": "integer"
          },
          "companyTitle": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "CommonInvoiceSummaryResponse": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "id",
          "accountId",
          "accountName",
          "title",
          "token",
          "publicUrl",
          "status",
          "totalOrders",
          "readyOrders",
          "paidOrders",
          "amount",
          "paid",
          "remaining",
          "amountKopecks",
          "paidKopecks",
          "remainingKopecks",
          "sentAt",
          "lastReminderAt",
          "nextReminderAt",
          "closedAt",
          "closedBy",
          "closeReason",
          "lastError",
          "paymentSuccessNotificationError",
          "tbankOrderId",
          "tbankPaymentId",
          "tbankPaymentAmountKopecks",
          "tbankTerminalLabel",
          "tbankTerminalKey",
          "paymentRouteType",
          "paymentRouteProvider",
          "paymentRouteProfileName",
          "paymentRouteRecipient",
          "paymentRouteManualTaskId",
          "contractorPaymentRoute",
          "paymentRouteSelectedAt",
          "invoicePurpose",
          "supersedesInvoiceId",
          "invoicePaymentMode",
          "paperInvoiceIssuedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "accountId": {
            "type": "integer"
          },
          "accountName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "publicUrl": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "totalOrders": {
            "type": "integer"
          },
          "readyOrders": {
            "type": "integer"
          },
          "paidOrders": {
            "type": "integer"
          },
          "amount": {
            "type": "number"
          },
          "paid": {
            "type": "number"
          },
          "remaining": {
            "type": "number"
          },
          "amountKopecks": {
            "type": "integer"
          },
          "paidKopecks": {
            "type": "integer"
          },
          "remainingKopecks": {
            "type": "integer"
          },
          "sentAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "lastReminderAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "nextReminderAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "closedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "closedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "closeReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "lastError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "paymentSuccessNotificationError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "tbankOrderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "tbankPaymentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "tbankPaymentAmountKopecks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "tbankTerminalLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "tbankTerminalKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "paymentRouteType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "paymentRouteProvider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "paymentRouteProfileName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "paymentRouteRecipient": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "paymentRouteManualTaskId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "contractorPaymentRoute": {
            "type": "boolean"
          },
          "paymentRouteSelectedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoicePurpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "supersedesInvoiceId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoicePaymentMode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "paperInvoiceIssuedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "CommonBillingAccountResponse": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "id",
          "name",
          "enabled",
          "autoRepeatOrders",
          "managerId",
          "managerName",
          "invoiceCompanyId",
          "invoiceCompanyTitle",
          "companies",
          "currentInvoice",
          "invoicePaymentMode"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "autoRepeatOrders": {
            "type": "boolean"
          },
          "managerId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "managerName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoiceCompanyId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoiceCompanyTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "companies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommonBillingCompanyResponse"
            }
          },
          "currentInvoice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommonInvoiceSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoicePaymentMode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
    }
  },
...
}